♻️ refactor: split MockLLMService block-gate to sibling file
MockLLMService.swift sat at exactly 400 lines — the swiftlint
`file_length` cap, enforced as an error under --strict. The next
addition would trip the cap as an unrelated lint failure (#726).

Carve the signal-blocked-gate test-helper cluster (the `BlockGate`
enum + blockGenerateUntilSignal / unblockGenerate /
awaitBlockReleaseIfArmed / extractParkedBlockContinuation) into
`MockLLMService+BlockGate.swift`, following the Type+Feature.swift
convention. Body file drops to 293 lines.

`state` and `State` widen private → internal (LLM-module-internal
only; no public surface change) so the sibling can drive the lock
and read `State.blockGate`. The sibling is a `nonisolated extension`
(swift-isolation Pattern 3 + Pattern 1: awaitBlockReleaseIfArmed
builds escaping continuation closures).

Closes #726
